requirements.txt
requests
mysql-connector
lxml
bs4